fix(x86): enable real SIMD on GCC and fix critical memory/scheduler bugs
Fixes and optimizations for the x86/GCC path. Benchmarks were run against a
clean build of the previous HEAD (46582f6) with alternating runs and medians
on a Ryzen 5 3600 (6c/12t) with GCC 15.2, -O3 -DNDEBUG -march=native.

SIMD / performance
------------------
* NativeSimd: enable GCC vector extensions for GCC >= 9 (were Clang-only).
  On GCC the class fell back to the unrolled array representation and real
  recurrences compiled to scalar chains (vfmadd231ss/vmulss with stack
  spills). Packed SIMD is now emitted (vfmadd231ps/vmulps/vdivps):
  AudioBiquad P=4 goes from 77.8 ms to 41.9 ms (2.02x -> 3.75x vs scalar).
  The VecType converting constructor becomes a constructor template (GCC
  rejects the non-template overload) and the scalar broadcast uses vector
  arithmetic to avoid -Wmaybe-uninitialized. Users can force the old path
  with -DLEGOLAS_HAS_VECTOR_EXTENSIONS=0.

* Array: add a serial cutoff (LEGOLAS_PARALLEL_THRESHOLD, default 32768
  elements) for fill/assign so small buffers no longer pay a thread
  wake-up, and deduplicate the chunked memcpy/fill paths.

* Examples: select the pack size from the hardware width (16 AVX-512,
  8 AVX2, 4 otherwise) instead of hardcoding P=4. OptionPricing P=8 goes
  from 177.8 ms to 94.9 ms (3.75x -> 7.04x), parallel 20.6 -> 13.2 ms.
  DepthwiseConv would otherwise regress with vector extensions because GCC
  already auto-vectorizes its spatial loop with the scalar fallback.

* Reductions: accumulate one vector across the whole loop in
  dotAssumeZeroPadding instead of doing a horizontal reduction per pack;
  process all full packs vectorially and only the partial tail scalar (the
  previous code always handled the last pack scalar and underflowed on
  empty arrays). Fix packed accumulator element access (operator[] instead
  of operator()) so dot/squaredNorm actually compile for packed arrays -
  that code path had never been instantiated.

Correctness
-----------
* Array: the defaulted move constructor left owner_/dataPtr_ in the
  moved-from object, causing a double free (reproduced with std::move).
  Implement a real move constructor that nulls the source. Add a
  static_assert requiring a trivially copyable RealType since the aligned
  allocator never constructs/destroys elements and bulk copies use memcpy.

* StaticScheduler: nested parallel_for corrupted the single current work
  item (stack use-after-free). A thread_local in_parallel_region() flag now
  runs nested regions inline, and a dispatch mutex serializes regions
  issued by distinct user threads. Read LEGOLAS_NUM_THREADS / OMP_NUM_THREADS
  when building the pool (legolas_thread_number() was never called) and make
  the spin count configurable (LEGOLAS_SPIN_COUNT).

* ImageFilter: no longer allocates multi-GB 4D arrays (capped at nx=320 with
  bad_alloc handling) and no longer segfaults when lena.bmp is not found
  from the current directory.

Build / hygiene
---------------
* CMake: Debug is no longer compiled with -O3; -fopt-info-vec-all,
  -march=native and verbose makefiles are opt-in options; add
  LEGOLAS_INTERFACE_OPTIMIZE to propagate optimization flags to
  FetchContent consumers.

* StaticArray: fix dead RecursiveFunctions::copy calling a missing load();
  replace invalid pragmas (clang loop, vector aligned, ivdep, assume) with a
  portable LEGOLAS_IVDEP macro. Remove unused typedefs and sign-compare
  warnings in tests, and drop the dead WorkStealing include.

* MultiThomasExample: warm up and report the best of 5 runs instead of a
  single cold measurement.

Validation: 7/7 CTest pass, 0 compiler warnings (was 312), packed SIMD
confirmed by disassembly, dedicated reproducers for the move and reduction
fixes. The large benchmarks are DRAM-bandwidth-bound on this host (~25 GB/s
measured), which caps parallel scaling; the scheduler itself scales 11.1x on
12 threads on a compute-only workload.

+++ tst/MultiPrec/MultiPrec.cxx
@@ -34,7 +34,7 @@ template <class V1L, class V1R>
 void convert(const V1L & v1l, V1R & v1r){
   assert(v1l.size()==v1r.size());
   using RealType=typename V1R::RealType;
-  for (size_t i=0 ; i<v1r.size() ; i++){
+  for (int i=0 ; i<v1r.size() ; i++){
     v1r[i]=RealType(v1l[i]);
   }
 }
